Moved Map to MapComponent for reuse accross the tour. Need to fix the spacing so that all elements fit on one page
